rename files and modify configs
