+++ .gitignore
@@ -27,3 +27,6 @@ node_modules
 
 # local secrets
 .env
+
+# editors
+.vscode
